refactor: split views
